requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
easyocr==1.7.2
Pillow==11.1.0
PyMuPDF==1.25.3
pdfplumber==0.11.4
numpy==1.26.4
tabulate==0.9.0
pytest==8.3.4
httpx==0.28.1
python-dotenv==1.0.0
groq>=0.28.0
langfuse>=2.0.0
paddleocr[doc-parser]>=3.2.0
unstructured[pdf]>=0.16.0